+++ node/resolvers/orderForm.ts
@@ -1,8 +1,28 @@
-import { prop, propOr } from 'ramda'
+import { prop, propOr, compose, forEach } from 'ramda'
 
+import { CHECKOUT_COOKIE, parseCookie } from '../utils'
 import { adjustItems } from './items'
 import { fillMessages } from './messages'
 import { getShippingInfo } from './shipping/utils/shipping'
+
+interface StoreSettings {
+  enableOrderFormOptimization: boolean
+  storeName: string
+  titleTag: string
+  metaTagDescription: string
+  metaTagKeywords: string
+  enableCriticalCSS: boolean
+}
+
+const SetCookieWhitelist = [CHECKOUT_COOKIE, '.ASPXAUTH']
+
+const isWhitelistedSetCookie = (cookie: string) => {
+  const [key] = cookie.split('=')
+  return SetCookieWhitelist.includes(key)
+}
+
+const replaceDomain = (host: string) => (cookie: string) =>
+  cookie.replace(/domain=.+?(;|$)/, `domain=${host};`)
 
 export const root = {
   OrderForm: {
@@ -44,6 +64,22 @@ export const root = {
   },
 }
 
+export async function forwardCheckoutCookies(
+  rawHeaders: Record<string, any>,
+  ctx: Context
+) {
+  const responseSetCookies: string[] = rawHeaders?.['set-cookie'] || []
+
+  const host = ctx.get('x-forwarded-host')
+  const forwardedSetCookies = responseSetCookies.filter(isWhitelistedSetCookie)
+  const parseAndClean = compose(parseCookie, replaceDomain(host))
+  const cleanCookies = forwardedSetCookies.map(parseAndClean)
+  forEach(
+    ({ name, value, options }) => ctx.cookies.set(name, value, options),
+    cleanCookies
+  )
+}
+
 export const queries = {
   orderForm: async (
     _: unknown,
@@ -52,7 +88,23 @@ export const queries = {
   ): Promise<CheckoutOrderForm> => {
     const { clients } = ctx
 
-    const newOrderForm = await clients.checkout.orderForm()
+    const {
+      data: newOrderForm,
+      headers,
+    } = await clients.checkout.orderFormRaw()
+
+    /**
+     * In case the enableOrderFormOptimization setting is enabled in the store,
+     * this will be the only `orderForm` query performed in the client. So no
+     * 'checkout.vtex.com' cookie would have been set. Otherwise vtex.store-graphql
+     * would've already set this cookie and this resolver should not overwrite it.
+     */
+    const storeSettings: StoreSettings = await clients.apps.getAppSettings(
+      'vtex.store@2.x'
+    )
+    if (storeSettings.enableOrderFormOptimization) {
+      forwardCheckoutCookies(headers, ctx)
+    }
 
     return newOrderForm
   },