+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `orderForm` query did not forward set-cookies from Checkout API, causing `checkout.vtex.com` cookie to never be set.
 
 ## [0.29.0] - 2020-03-25
 ### Added